--- ExemploServletJSTL/src/main/java/controller/PessoaServlet.java
@@ -1,32 +1,0 @@
-package controller;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import model.Pessoa;
-
-@WebServlet("/pessoa")
-public class PessoaServlet extends HttpServlet {
-	private static final long serialVersionUID = 1L;
-       
-    public PessoaServlet() {
-        super();
-    }
-
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		//System.out.println("Entrou no servlet PessoaServlet");
-		String idParam = request.getParameter("id");
-		String nomeParam = request.getParameter("nome");
-		
-		Pessoa pessoa = new Pessoa();
-		pessoa.setId(Integer.parseInt(idParam));
-		pessoa.setNome(nomeParam);
-		
-		System.out.println(pessoa);
-	}
-
-}
